Update int itc_countWords.cpp

+++ int itc_countWords.cpp
@@ -1,8 +1,8 @@
 #include "middle_str.h"
 int itc_countWords(string a){
-    long long b = 0, c = 0;
+    int b = 0, c = 0;
     string f = "";
-    for (long long i = 0; str[i] != '\0'; i++){
+    for (int i = 0; str[i] != '\0'; i++){
         if ((65 <= a[i] and a[i] <= 90) or (97 <= a[i] and a[i] <= 122)){
             f = f + str[i];
             c++;
